refactor: entity -> remote 패키지 아래로 이동
