+++ android/app/src/main/java/com/woowacourse/ody/data/remote/repository/DefaultMeetingRepository.kt
@@ -1,8 +1,8 @@
 package com.woowacourse.ody.data.remote.repository
 
-import com.woowacourse.ody.data.entity.meeting.MeetingRequest
-import com.woowacourse.ody.data.entity.meeting.MeetingResponse
-import com.woowacourse.ody.data.entity.toMeeting
+import com.woowacourse.ody.data.remote.entity.meeting.MeetingRequest
+import com.woowacourse.ody.data.remote.entity.meeting.MeetingResponse
+import com.woowacourse.ody.data.remote.entity.toMeeting
 import com.woowacourse.ody.data.remote.RetrofitClient
 import com.woowacourse.ody.data.remote.service.MeetingService
 import com.woowacourse.ody.domain.model.Meeting
