+++ android/app/src/main/java/com/woowacourse/ody/data/remote/service/JoinService.kt
@@ -1,7 +1,7 @@
 package com.woowacourse.ody.data.remote.service
 
-import com.woowacourse.ody.data.entity.join.JoinRequest
-import com.woowacourse.ody.data.entity.meeting.MeetingResponse
+import com.woowacourse.ody.data.remote.entity.join.JoinRequest
+import com.woowacourse.ody.data.remote.entity.meeting.MeetingResponse
 import retrofit2.http.Body
 import retrofit2.http.POST
