+++ android/app/src/main/java/com/woowacourse/ody/domain/repository/JoinRepository.kt
@@ -1,7 +1,7 @@
 package com.woowacourse.ody.domain.repository
 
-import com.woowacourse.ody.data.entity.join.JoinRequest
-import com.woowacourse.ody.data.entity.meeting.MeetingResponse
+import com.woowacourse.ody.data.remote.entity.join.JoinRequest
+import com.woowacourse.ody.data.remote.entity.meeting.MeetingResponse
 
 interface JoinRepository {
     suspend fun postMates(joinRequest: JoinRequest): Result<MeetingResponse>
